Move test_data to main crate
