+++ src/test/fixtures/v10.rs
@@ -1,5 +1,5 @@
-use super::*;
-use cloudevents::{Event, EventBuilder, EventBuilderV10};
+use crate::test::fixtures::*;
+use crate::{Event, EventBuilder, EventBuilderV10};
 use serde_json::{json, Value};
 use url::Url;
