--- tests/test_data/mod.rs
@@ -1,8 +1,0 @@
-#[allow(dead_code)]
-mod data;
-pub use data::*;
-
-#[allow(dead_code)]
-pub mod v03;
-#[allow(dead_code)]
-pub mod v10;
